third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/debugger-uncaught-promise-on-pause-expected.txt

CONSOLE MESSAGE: line 10: console.clear
CONSOLE ERROR: line 16: Uncaught (in promise) Error: inspector.err1
CONSOLE ERROR: line 34: Uncaught (in promise) Error: inspector.err2
Tests uncaught promise rejections fired during pause.

Set timer for test function.
debugger-uncaught-promise-on-pause.html:10 Console was cleared
debugger-uncaught-promise-on-pause.html:16 Uncaught (in promise) Error: inspector.err1
runPromises @ debugger-uncaught-promise-on-pause.html:16
timeout @ debugger-uncaught-promise-on-pause.html:41
debugger-uncaught-promise-on-pause.html:34 Uncaught (in promise) Error: inspector.err2
runPromises @ debugger-uncaught-promise-on-pause.html:34
timeout @ debugger-uncaught-promise-on-pause.html:41
